# src/config/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Bookstore-Api",
        "description": "Bookstore-App Api documentation.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/books/titles/{slug}": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Get complete title by slug",
                "description": "Retrieves complete title information by its unique identifier",
                "operationId": "getTitleById",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Unique slug identifier of the title",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Title not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}